added RoomType table 	- defining room types (replacing the current hard coded ones) 	- room types are academic session independent added RoomTypeOption table 	- session dependent 	- relation (RoomType, Session) -> status (set of booleans), message 	- added status canScheduleEvents 		- use Room.getRoomType().getOption(Session).canScheduleEvents() to check whether it is possible to schedule events into a room of given type 	- these options (and messages) are editable via Edit Academic Session page

+++ JavaSource/org/unitime/timetable/solver/SolverProxy.java
@@ -101,7 +101,7 @@ public interface SolverProxy extends ClassAssignmentProxy {
 	public String getHostLabel();
 	public void dispose() throws Exception;
 	
-	public RoomReport getRoomReport(int startDay, int endDay, int nrWeeks, Integer roomType) throws Exception;
+	public RoomReport getRoomReport(int startDay, int endDay, int nrWeeks, Long roomType) throws Exception;
 	public DeptBalancingReport getDeptBalancingReport() throws Exception;
 	public ViolatedDistrPreferencesReport getViolatedDistrPreferencesReport() throws Exception;
 	public DiscouragedInstructorBtbReport getDiscouragedInstructorBtbReport() throws Exception;
